Renaming the file containing SubmissionBuilder to submission_builder.py
